# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "L5 Swagger OpenApi description",
        "version": "1.0.0"
    },
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login",
                "operationId": "login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "anhhn@vnext.vn"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/api/logout": {
            "delete": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout",
                "operationId": "Logout",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "bearer"
                                ],
                                "properties": {
                                    "bearer": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/change-pass/{id}": {
            "put": {
                "tags": [
                    "Auth"
                ],
                "summary": "ChangePass",
                "operationId": "ChangePass",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id",
                                    "bearer",
                                    "old_password",
                                    "new_password",
                                    "new_password_confirmation"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "string",
                                        "example": "1"
                                    },
                                    "bearer": {
                                        "type": "string"
                                    },
                                    "old_password": {
                                        "type": "string",
                                        "example": "123456"
                                    },
                                    "new_password": {
                                        "type": "string",
                                        "example": "123456"
                                    },
                                    "new_password_confirmation": {}
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/members/edit/{id}": {
            "get": {
                "tags": [
                    "Members"
                ],
                "summary": "Detail members",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/members/update/1": {
            "put": {
                "tags": [
                    "Members"
                ],
                "summary": "Updates a member",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "avatar",
                        "in": "query",
                        "schema": {
                            "type": "file"
                        }
                    },
                    {
                        "name": "bearer",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "avatar_official",
                        "in": "query",
                        "schema": {
                            "type": "file"
                        }
                    },
                    {
                        "name": "gender",
                        "in": "query",
                        "description": "Gender",
                        "schema": {
                            "type": "radio"
                        },
                        "examples": {
                            "int": {
                                "summary": "male",
                                "value": "1"
                            },
                            "uuid": {
                                "summary": "female",
                                "value": "2"
                            }
                        }
                    },
                    {
                        "name": "nick_name",
                        "in": "query",
                        "schema": {
                            "type": "text",
                            "example": "long"
                        }
                    },
                    {
                        "name": "birth_date",
                        "in": "query",
                        "schema": {
                            "type": "date",
                            "example": "1999/02/02"
                        }
                    },
                    {
                        "name": "other_email",
                        "in": "query",
                        "schema": {
                            "type": "email",
                            "example": "long@gmail.com"
                        }
                    },
                    {
                        "name": "identity_number",
                        "in": "query",
                        "schema": {
                            "type": "number",
                            "example": "1234567899"
                        }
                    },
                    {
                        "name": "identity_card_date",
                        "in": "query",
                        "schema": {
                            "type": "date",
                            "example": "2000/12/02"
                        }
                    },
                    {
                        "name": "identity_card_place",
                        "in": "query",
                        "schema": {
                            "type": "date",
                            "example": "Việt Nam"
                        }
                    },
                    {
                        "name": "skype",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "example": "longtt"
                        }
                    },
                    {
                        "name": "facebook",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "example": "Long Bầu"
                        }
                    },
                    {
                        "name": "passport_number",
                        "in": "query",
                        "schema": {
                            "type": "number",
                            "example": "213123123123123"
                        }
                    },
                    {
                        "name": "passport_expiration",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "example": "2022/01/01"
                        }
                    },
                    {
                        "name": "nationality",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "example": "Việt Nam"
                        }
                    },
                    {
                        "name": "bank_name",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "example": "Techcombank"
                        }
                    },
                    {
                        "name": "bank_account",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "example": "longtt"
                        }
                    },
                    {
                        "name": "marital_status",
                        "in": "query",
                        "schema": {
                            "type": "number",
                            "example": "1"
                        }
                    },
                    {
                        "name": "academic_level",
                        "in": "query",
                        "schema": {
                            "type": "date",
                            "example": "12/12"
                        }
                    },
                    {
                        "name": "permanent_address",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "example": "Hạ Bằng - Thạch Thất - Hà Nội"
                        }
                    },
                    {
                        "name": "temporary_address",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "example": "32/80 - Đỗ Đức Dục - Mễ Trì - Hà Nội"
                        }
                    },
                    {
                        "name": "tax_identification",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "example": "0123456789"
                        }
                    },
                    {
                        "name": "healthcare_provider",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "example": "Hà Nội"
                        }
                    },
                    {
                        "name": "emergency_contact_name",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "example": "Phương Anh"
                        }
                    },
                    {
                        "name": "emergency_contact_relationship",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "example": "0359146002"
                        }
                    },
                    {
                        "name": "emergency_contact_number",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "example": "0359146004"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/register-ot/{id}": {
            "get": {
                "tags": [
                    "Register overtime"
                ],
                "summary": "Detail register overtime",
                "operationId": "create",
                "parameters": [
                    {
                        "name": "bearer",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Register overtime"
                ],
                "summary": "Create register overtime",
                "operationId": "store",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "bearer",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "request_ot_time",
                        "in": "query",
                        "schema": {
                            "type": "date"
                        }
                    },
                    {
                        "name": "reason",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/register-ot/edit/{id}": {
            "put": {
                "tags": [
                    "Register overtime"
                ],
                "summary": "Edit register overtime",
                "operationId": "updateRegisterOT",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "bearer",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "request_ot_time",
                        "in": "query",
                        "schema": {
                            "type": "date"
                        }
                    },
                    {
                        "name": "reason",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}